requirements.txt
streamlit
pydantic
pydantic-ai
python-dotenv
httpx
pandas
openai